# Precompilation support, consistent speed improvements!