--- a/Project.toml
+++ b/Project.toml
@@ -5,10 +5,12 @@ version = "0.1.1"
 [deps]
 ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
 JpegTurbo = "b835a17e-a41a-41e7-81f0-2f016b05efe0"
+SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
 libexif_jll = "cdeeb48b-bcdf-5b3f-98c4-7a29487f695f"
 
 [compat]
-julia = "1.7"
-libexif_jll = "0.6"
 ColorTypes = "0.11.4"
 JpegTurbo = "0.1.1"
+SnoopPrecompile = "1"
+julia = "1.7"
+libexif_jll = "0.6"