Following the path of https://github.com/JuliaImages/ImageFiltering.jl/pull/255 and   https://github.com/JuliaImages/ImageQualityIndexes.jl/pull/54 , Use of precompilation and utilization of latest options in precompilation through use of pkgimage has some great effects and improvements all across the board.

For me, initial invocation of 
```jl
@time @eval read_tags("tmp.jpg"; read_all = true)
```
improves with latest options

Combination of precompilation and latest options in precompilation
In 1.9.0 dropped from 0.58 seconds to 0.015 seconds, approximately 40x improvements which is impressive.
In 1.8.5 dropped from 0.55 seconds to 0.37 seconds on use of precompilation.

I saw improvements almost all across board in benchmarks that can be found in gist below. One concern though for me is size of compiled caches of ExifViewer even in my basic testing, in 1.8.5 it was 1.92 mb(only .ji file) while in 1.9.0 it was 5.04mb(.ji + .dll file). ExifViewer.jl is a small package with just 2 functions and yet taking a lot of space surprisingly.

More details can be found here: https://gist.github.com/ashwani-rathee/d250abdd0172603448b0e81a96815681